+++ src/Repository/SeasonRepository.php
@@ -2,8 +2,10 @@
 
 namespace App\Repository;
 
+use App\Entity\Race;
 use App\Entity\Season;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\PersistentCollection;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
@@ -19,6 +21,33 @@ class SeasonRepository extends ServiceEntityRepository
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, Season::class);
+    }
+
+    public function findSeasonWithSingleRaceDataForScores(int $seasonId, Race $race): ?Season
+    {
+        /** @var Season $season */
+        $season = $this->createQueryBuilder('s')
+            ->join('s.races', 'r')
+            ->join('r.raceResults', 'rr')
+            ->join('r.raceResultBets', 'rrb')
+            ->join('r.penaltyPointsAwards', 'ppa')
+            ->where('s.id = :seasonId')
+            ->andWhere('r.id = :raceId')
+            ->andWhere('rr.race = :race')
+            ->andWhere('rrb.race = :race')
+            ->andWhere('ppa.race = :race')
+            ->setParameter('seasonId', $seasonId)
+            ->setParameter('raceId', $race->getId())
+            ->setParameter('race', $race)
+            ->addSelect('r')
+            ->addSelect('rr')
+            ->addSelect('rrb')
+            ->addSelect('ppa')
+            ->getQuery()
+            ->getOneOrNullResult()
+        ;
+
+        return $season;
     }
 
     //    /**